Removal of SVG export
This resolves #9. Frankly, the SVG files are too large and with the new ability to export 350dpi tiff files, it's not necessary in this application

+++ gui/dropletgraphwidget.cpp
@@ -210,7 +210,7 @@ void DropletGraphWidget::setConvexHullsVisible(bool b)
 
 void DropletGraphWidget::updateConvexHullCount()
 {
-    if (m_data->colorComponentCount() > m_convexHulls.size()) {
+    if (m_data->colorComponentCount() > (size_t)m_convexHulls.size()) {
         for (size_t i = m_convexHulls.size() + 1; i < m_data->colorComponentCount() ; ++i) {
             Plot::Polygon * p = new Plot::Polygon(horizontalAxis(), verticalAxis());
             p->setAutoColor(false);
